refactor(agents): delete public wrapAdapterAsManager export, add fakeAgentManager test helper (Tasks 7-10)
