+++ src/tdd/session-runner.ts
@@ -5,7 +5,7 @@
  */
 
 import type { AgentAdapter } from "../agents";
-import { resolveDefaultAgent, wrapAdapterAsManager } from "../agents";
+import { resolveDefaultAgent } from "../agents";
 import type { ModelTier, NaxConfig } from "../config";
 import { resolveModelForAgent } from "../config";
 import { createContextToolRuntime } from "../context/engine";
@@ -124,6 +124,7 @@ export interface TddSessionBinding {
 export async function runTddSession(
   role: TddSessionRole,
   agent: AgentAdapter,
+  agentManager: import("../agents/manager-types").IAgentManager,
   story: UserStory,
   config: NaxConfig,
   workdir: string,
@@ -246,11 +247,8 @@ export async function runTddSession(
   // Run the agent. When a sessionBinding is provided, route through
   // SessionManager.runInSession so state transitions (CREATED → RUNNING →
   // COMPLETED/FAILED) and bindHandle happen automatically (#541, #589).
-  // ADR-013 Phase 1: runInSession now takes IAgentManager. Use the binding's
-  // agentManager when present; otherwise wrap adapter for a no-fallback path.
-  // Absent binding path stays compatible with tests that skip SessionManager.
   const effectiveManager: import("../agents/manager-types").IAgentManager =
-    sessionBinding?.agentManager ?? wrapAdapterAsManager(agent);
+    sessionBinding?.agentManager ?? agentManager;
 
   const result = sessionBinding
     ? await sessionBinding.sessionManager.runInSession(sessionBinding.sessionId, effectiveManager, {
